Add first element with Cart, Actions and Make an Order options to menu
