+++ helpers/fb_chat_replying.py
@@ -1,0 +1,91 @@
+import requests
+
+from api.moltin_requests import fetch_products
+from helpers.fb_items_formatters import (
+    get_first_menu_element,
+    get_menu_element,
+)
+
+
+def send_menu(recipient_id, auth_token, fb_token):
+    params = {'access_token': fb_token}
+    headers = {'Content-Type': 'application/json'}
+
+    products = fetch_products(auth_token)
+
+    first_menu_element = [get_first_menu_element()]
+    elements = [
+        get_menu_element(product, auth_token)
+        for product in products['data'][:5]
+    ]
+    elements = [*first_menu_element, *elements]
+
+    request_content = {
+        'recipient': {'id': recipient_id},
+        'message': {
+            'attachment': {
+                'type': 'template',
+                'payload': {
+                    'template_type': 'generic',
+                    'elements': elements,
+                },
+            },
+        },
+    }
+
+    response = requests.post(
+        'https://graph.facebook.com/v2.6/me/messages',
+        headers=headers,
+        params=params,
+        json=request_content,
+    )
+    response.raise_for_status()
+
+
+def send_button(recipient_id, fb_token):
+    params = {'access_token': fb_token}
+    headers = {'Content-Type': 'application/json'}
+
+    request_content = {
+        'recipient': {'id': recipient_id},
+        'message': {
+            'attachment': {
+                'type': 'template',
+                'payload': {
+                    'template_type': 'button',
+                    'text': 'Try the postback button!',
+                    'buttons': [
+                        {
+                            'type': 'postback',
+                            'title': 'Postback Button',
+                            'payload': 'DEVELOPER_DEFINED_PAYLOAD',
+                        }
+                    ],
+                },
+            }
+        },
+    }
+
+    response = requests.post(
+        'https://graph.facebook.com/v2.6/me/messages',
+        headers=headers,
+        params=params,
+        json=request_content,
+    )
+    response.raise_for_status()
+
+
+def send_message(recipient_id, message_text, fb_token):
+    params = {'access_token': fb_token}
+    headers = {'Content-Type': 'application/json'}
+    request_content = {
+        'recipient': {'id': recipient_id},
+        'message': {'text': message_text},
+    }
+    response = requests.post(
+        'https://graph.facebook.com/v2.6/me/messages',
+        params=params,
+        headers=headers,
+        json=request_content,
+    )
+    response.raise_for_status()
